Move functionality to its own feature.
